Gravity and TopBottom Collision FInished!
Start to add formations and fix horizontal collision to for a sidescroll loss. Good work!

+++ src/GameObject.java
@@ -2,8 +2,8 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public class GameObject {
-	int x;
-	int y;
+	double x;
+	double y;
 	int width;
 	int height;
 	boolean active;
@@ -17,6 +17,6 @@ public class GameObject {
 		colBox = new Rectangle(x,y,width,height);
 	}
 	void setBounds() {
-		colBox = new Rectangle(x,y,width,height);
+		colBox = new Rectangle((int) x,(int) y,width,height);
 	}
 }
